change files to fit the check style configuration
